Move Int32Tests
